+++ src/main/java/org/complete/challang/drink/service/DrinkService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.complete.challang.common.exception.ApiException;
 import org.complete.challang.common.exception.ErrorCode;
 import org.complete.challang.drink.domain.entity.Drink;
-import org.complete.challang.drink.repository.DrinkFindResponse;
+import org.complete.challang.drink.dto.response.DrinkFindResponse;
 import org.complete.challang.drink.repository.DrinkRepository;
 import org.springframework.stereotype.Service;
