+++ src/test/java/org/complete/challang/drink/controller/DrinkControllerTest.java
@@ -7,7 +7,7 @@ import org.complete.challang.common.exception.ErrorCode;
 import org.complete.challang.common.exception.GlobalExceptionHandler;
 import org.complete.challang.drink.common.adaptor.LocalDateTimeTypeAdaptor;
 import org.complete.challang.drink.domain.entity.Drink;
-import org.complete.challang.drink.repository.DrinkFindResponse;
+import org.complete.challang.drink.dto.response.DrinkFindResponse;
 import org.complete.challang.drink.service.DrinkService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
